+++ lib/rubocop-thread_safety.rb
@@ -8,6 +8,7 @@ require 'rubocop/thread_safety/plugin'
 
 require 'rubocop/cop/mixin/operation_with_threadsafe_result'
 
+require 'rubocop/cop/thread_safety/active_support_callbacks'
 require 'rubocop/cop/thread_safety/class_instance_variable'
 require 'rubocop/cop/thread_safety/class_and_module_attributes'
 require 'rubocop/cop/thread_safety/mutable_class_instance_variable'